salary changed from long to double

+++ practise/src/practisegit/Employee.java
@@ -5,7 +5,7 @@ public class Employee {
 	private String name;
 	private String state;
 	private String address; // Conflicts resolved
-	private long salary;	
+	private double salary;	
 	private int age;
 
 }
